Emitter on enemy; Broke things.
I somehow managed to break the simulated iPhone, and it stopped responding to any touch commands.  It didn't even call the touch function, and I don't know why. I force quited the simulator, then ran again, and it worked. I don't know why it's happening, but I feel it's a sign of "It's time to stop".